raw data update
Fixed the raw data transfer to model
